+++ react-app/src/components/Dashboard.js
@@ -4,7 +4,6 @@ import DashNav from "./DashNav";
 import HistoryTab from "./HistoryTab";
 import FriendsTab from "./FriendsTab";
 import BillsTab from "./BillsTab";
-import ExpenseBill from "./BillsTab/ExpenseBill";
 
 import './Dashboard.css'
 
@@ -23,9 +22,6 @@ const Dashboard = () => {
 					</Route>
 					<Route path="/transaction-history">
 						<HistoryTab />
-					</Route>
-					<Route path='/expenses/:expenseId/bill'>
-						<ExpenseBill />
 					</Route>
 				</Switch>
 			</div>
